+++ changelog.md
@@ -1,3 +1,10 @@
+# Changelog
+
+### Version 1.0.3
+- Add GitHub emoji counter
+- Add original Unicode
+- Add JSON version
+
 ### Version 1.0.2
 - Add GitHub Actions badge
 - Fix minor visual bug of the markdown